+++ packages/ui/src/components/Toasts/Snackbar.tsx
@@ -1,0 +1,32 @@
+import { Stack, styled } from '@mui/material'
+import { useToasts } from '../../contexts/ToastContext'
+import ToastBar from './ToastBar'
+
+interface Props {
+  className?: string
+}
+
+const Snackbar = ({ className }: Props) => {
+  const { toasts } = useToasts()
+
+  return (
+    <Stack
+      className={className}
+      flexDirection="column"
+      gap={1}
+    >
+      {toasts.map((toast) => (
+        <ToastBar
+          toast={toast}
+          key={toast.id}
+        />
+      ))}
+    </Stack>
+  )
+}
+
+export default styled(Snackbar)`
+  position: absolute;
+  bottom: 1rem;
+  left: 1rem;
+`
